fix: minor bug
If user hovers Submit Data waiting for VoS then votes it won't update but will try to submit a vote for the last VoS which will fail and waste their vote.. With the other changes that have been made to mostly automate things we should be fine to fetch the clan data on click instead of mouseenter now.

I guess I'll find out if this breaks anything soon enough - but it should be fine.