fix: add redundant voting check (for good reason) ; remove redundant validity check
Votes are cleared near the hour so if our vote is >5 minutes old then we have already voted regardless if Voted = False which can sometimes occur due to how we clear 'Voted' periodically to allow for multiple votes during primetime but also to avoid someone being unable to vote

In addition we already have a check for "!hasValidData()" which is an early return before we submit data. this means that we do not need to check it again prior to voting.